backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.7.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
